some of ReferenceCollectionReaderWriter needed to still be public
