--- source/Nevermore/Mapping/IInheritedClassSerialization.cs
@@ -1,6 +1,0 @@
-namespace Nevermore.Mapping
-{
-    public interface IInheritedClassSerialization
-    {
-    }
-}
