+++ source/Nevermore/Mapping/IInheritedCustomTypeSerialization.cs
@@ -1,0 +1,6 @@
+namespace Nevermore.Mapping
+{
+    public interface IInheritedCustomTypeSerialization
+    {
+    }
+}
